lexflow-web/pyproject.toml
[project]
name = "lexflow-web"
version = "0.1.0"
description = "Web frontend for LexFlow workflow visualization and execution"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pyyaml>=6.0",
    "lexflow",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lexflow_web"]

[tool.hatch.build.targets.wheel.shared-data]
"src/lexflow_web/static" = "lexflow_web/static"
